# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: BiteBattle API
  version: 1.0.0
  description: |
    REST APIs for BiteBattle, a head-to-head restaurant selection app.
    
    **Authentication:**
    Most routes require an `Authorization` header with a Bearer token. Obtain this token by logging in via the `/v1/auth/login` route. Example:
    
        Authorization: Bearer <token>
    
    Only `/v1/auth/register` and `/v1/auth/login` do not require authentication.
...
paths:
...
  /v1/account:
    get:
      tags: [Account]
      summary: Get current user's profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
        '400':
          description: Invalid user id
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    put:
      tags: [Account]
      summary: Update profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProfileRequest' }
      responses:
        '200':
          description: Profile updated
        '400':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: Email already exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string }
        phone_number: { type: string, nullable: true }
        profile_pic_url: { type: string, format: uri, nullable: true }
        bio: { type: string, nullable: true }
        last_login_at: { type: string, format: date-time, nullable: true }
    UpdateProfileRequest:
      type: object
      required: [name, email]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        phone_number: { type: string, nullable: true }
        profile_pic_url: { type: string, format: uri, nullable: true }
        bio: { type: string, nullable: true }
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT